requirements.txt
jax
jaxlib
jaxtyping
# jax-metal
equinox
torch
torchvision
optax
tqdm
# tensorflow
# tensorflow-datasets
Pillow
ml-collections